Add and Tested Question and choices Endpoints

+++ ExaminationSystem.Application/Mediators/QuestionMediator.cs
@@ -1,4 +1,5 @@
-﻿using ExaminationSystem.Domain.Dtos.Question;
+﻿using ExaminationSystem.Application.Services;
+using ExaminationSystem.Domain.Dtos.Question;
 using ExaminationSystem.Domain.Mediators.contract;
 using ExaminationSystem.Domain.Services.contract;
 using System;
@@ -12,10 +13,12 @@ namespace ExaminationSystem.Application.Mediators
     public class QuestionMediator : IQuestionMediator
     {
         private readonly IQuestionService _questionService;
+        private readonly IExamQuestionService _examQuestionService;
 
-        public QuestionMediator(IQuestionService questionService)
+        public QuestionMediator(IQuestionService questionService, IExamQuestionService examQuestionService)
         {
             _questionService = questionService;
+            _examQuestionService = examQuestionService;
         }
 
         public async Task AddQuestion(QuestionCreateDto questionCreateDto)
@@ -25,7 +28,19 @@ namespace ExaminationSystem.Application.Mediators
 
         public async Task DeleteQuestion(int id)
         {
-            await _questionService.DeleteQuestion(id);
+            var Question = await _questionService.GetQuestionById(id);
+
+            if (Question != null)
+            {
+                await _questionService.DeleteQuestion(id);
+
+                var examQuestions = await _examQuestionService.Get(eq => eq.QuestionId == id);
+
+                if (examQuestions != null && examQuestions.Count() > 0)
+                {
+                    await _examQuestionService.DeleteRange(examQuestions);
+                }
+            }
         }
 
         public async Task<QuestionDto> GetQuestionById(int id)
